+++ db/migrate/20241004145138_add_creator_id_to_projects.rb
@@ -3,6 +3,6 @@ class AddCreatorIdToProjects < ActiveRecord::Migration[7.2]
 
     # Adds a project_creator_id column to the projects table
     # to associate a project with its creator.
-    add_column :projects, :project_creator_id, :integer
+    add_column :projects, :project_creator_id, :bigint
   end
 end
